Add files via upload
Sequence information parsing
In this exercise, you will create a program to “parse” sequence name information from a single line of a FASTQ formatted file. Your task is to create a
Python script called fastqParse that:
asks for and collects the seqname line of a FASTQ file using input()
parses out each field of the run information from the string and displays each of them on a new line For example, if I enter the FASTQ seqname line:
@EAS139:136:FC706VJ:2:2104:15343:197393 then the program will output:
Instrument = EAS139
Run ID = 136
Flow Cell ID = FC706VJ
Flow Cell Lane = 2
Tile Number = 2104
X-coord = 15343
Y-coord = 197393 ### Hints:
The input string is guaranteed to have 7 fields.
The first character of the FASTQ seqname line is “@” and each field of the run information is separated by a colon“:”.
A reasonable solution would be around 16 lines of code excluding comments.